Rename JukeboxArea.tsx to JukeboxAreaMusic.tsx
Having both JukeboxArea.ts and JukeboxArea.tsx have the same filename
made all of our imports ambiguous, which caused issues.
