+++ frontend/src/components/Town/interactables/JukeboxAreaMusic.tsx
